+++ flask_unchained/bundles/sqlalchemy/meta_options.py
@@ -1,5 +1,6 @@
 import inspect
 
+from flask_unchained import unchained
 from py_meta_utils import McsArgs, MetaOption
 from sqlalchemy.orm import RelationshipProperty
 from sqlalchemy_unchained import ModelMetaOptionsFactory as BaseModelMetaOptionsFactory
@@ -13,13 +14,19 @@ class ModelMetaOption(MetaOption):
     def __init__(self):
         super().__init__('model', default=None, inherit=True)
 
+    def get_value(self, Meta: Type[object], base_classes_meta, mcs_args: McsArgs):
+        value = super().get_value(Meta, base_classes_meta, mcs_args)
+        if value and unchained._models_initialized:
+            value = unchained.sqlalchemy_bundle.models[value.__name__]
+        return value
+
     def check_value(self, value, mcs_args: McsArgs):
         if mcs_args.Meta.abstract:
             return
 
         from .base_model import BaseModel
-        assert inspect.isclass(value) and issubclass(value, BaseModel), \
-            f'{mcs_args.name} is missing the model Meta attribute'
+        if not (inspect.isclass(value) and issubclass(value, BaseModel)):
+            raise TypeError(f'{mcs_args.name} is missing the model Meta attribute')
 
 
 class RelationshipsMetaOption(MetaOption):